requirements.txt
faust
langdetect
google-generativeai
praw
requests
confluent-kafka
pymongo
fastapi
uvicorn[standard]
websockets
python-multipart
pydantic
python-dotenv
aiohttp
requests
langchain-google-genai
langchain-core
langchain
kafka-python